[feature/swing] Refactoring: Klassen umbenannt
